ml/requirements.txt
pandas
scikit-learn
joblib
psycopg2-binary
python-dotenv
